\name{MPBDIAsolve}
\alias{MPBDIAsolve}
\title{
  Steady-state solution of microphytobenthos with C, N, P and O2 dynamics in the sediment.
}
\description{

  \code{MPBDIAsolve} finds the steady-state solution of the MPBDIA model (including microphytobenthos dynamics).
}
\usage{
  MPBDIAsolve(parms = list(), yini = NULL, Grid = NULL, porosity = NULL, 
     bioturbation = NULL, irrigation = NULL, diffusionfactor = NULL,  
     dynamicbottomwater = FALSE, ratefactor = NULL, verbose = FALSE,  
     times = c(0, 1e+06), method = "runsteady", ...)
}
\arguments{

  \item{parms}{ A list with parameter values, Available parameters can be listed using function 
  \link{MPBDIAparms}.

    See details.
  }

 \item{Grid}{ If specified: either an object, as returned by \code{setup.grid.1D} from the package \code{ReacTran}, 
   a vector of length 101 with the transport distances (from mid to mid of layers, upper interface = diffusive boundary layer), 
   or one number with the constant layer thickness. 
   If \code{NULL}, it is defined as \code{setup.grid.1D(x.up = 0, dx.1 = 0.01, N = 100, L = 100)}, i.e. the total length is 100 cm, 
   the first layer is 0.01 cm thick and layers are increasing with depth for 100 layers.  
  } 

 \item{porosity}{ If specified, either an object with porosities ([-]) as returned by \code{setup.prop.1D} from the package \code{ReacTran}, 
   a vector of length 101 with the porosities defined at the layer interfaces, or one number with the constant porosity. 
   If \code{NULL}, it is defined by the parameters \code{por0}, \code{pordeep} and \code{porcoeff} as: 
   \code{(pordeep + (por0 - pordeep) * exp(-pmax(0, x.int)/porcoeff))}, where \code{x.int} is the distance, from the surface of the layer interface. 
   Note that the porosity values should be consistent withe the \code{Grid} - and should be inbetween 0 and 1.
  }

 \item{bioturbation}{ If specified, either an object with bioturbation rates (units [cm2/d]) as returned by \code{setup.prop.1D} from the package \code{ReacTran}, 
   a vector of length 101 with the bioturbation defined at the layer interfaces, or one number with the constant bioturbation. 
   If \code{NULL}, it is defined by the parameters \code{biot}, \code{mixdepth} and \code{mixatt} as: 
   \code{biot * exp(-pmax(0, (x.int-mixL))/mixatt)}, where \code{x.int} is the distance, from the surface of the layer interface. 
   Note that the bioturbation values should be consistent withe the \code{Grid}.
  }

 \item{irrigation}{ If specified, either an object with irrigation rates (units [/d]) as returned by \code{setup.prop.1D} from the package \code{ReacTran}, 
   a vector of length 100 with the irrigation rates defined at the layer centres, or one number with the constant rates. 
   If \code{NULL}, it is defined by the parameters \code{irr}, \code{mixdepth} and \code{mixatt} as: 
   \code{irr * exp(-pmax(0, (x-mixdepth)/mixxatt))}, where \code{x} is the distance, from the surface, of the layer centres. 
   Note that the irrigation values should be consistent withe the \code{Grid}.
  }

 \item{diffusionfactor}{ The multiplication factor necessary to go from molecular diffusion to effective sediment diffusion, i.e. that takes into account tortuosity.
   If specified, either an object with these factors ([-]) as returned by \code{setup.prop.1D} from the package \code{ReacTran}, 
   a vector of length 101 with these factors defined at the layer interfaces, or one number with the constant factor. 
   If \code{NULL}, it is set equal to the porosity. Note that the factors should be consistent withe the \code{Grid}.
  }

  \item{yini }{Initial guess of the steady-state solution.
  }

 \item{dynamicbottomwater}{ If \code{TRUE}, then the concentrations in the water overlying the sediment will also be dynamically described, and with water height equal to \code{Hwater}. Note that this will slow down the simulation.   
  }
 \item{ratefactor}{ \code{NULL} or a list, detailing the forcing function for the  biogeochemical rate multiplication factor. If not specified (or \code{NULL}), then it is assumed to be 1 and constant. 
    If a \code{list}, it should contain either a data time series (\code{list(data = )}) or parameters determining the periodicity of the seasonal signal (defined as \code{list(data = NULL, amp = 0, period = 365, phase = 0, pow = 1, min = 0)}. see details. 
  }  
 \item{verbose}{ If TRUE, will write progession to the screen .
  }
   \item{times}{ the time span over which the model needs to run to find steady-state
  }
   \item{method}{ the steady-state solution method. 
  }
  \item{... }{Any argument passed to the steady-state solver.
  }
}

\value{
 \code{MPBDIAsolve} returns an object of class \code{MPBDIAstd}, and of class \code{steady1D}, as generated by the solvers from R-package \code{rootSolve} (\link{steady.1D}[rootSolve]).
 It contains the elements:
 \itemize{
   \item \code{y}, with the state variables at steady-state (\code{FDET, SDET, O2, NO3, NO2, NH3, ODU, PO4, FeP, CaP, DIC}).
   \item \code{O2flux, O2deepflux, NO3flux, NO3deepflux, NO2flux, NO2deepflux, NH3flux, NH3deepflux, 
      ODUflux, ODUdeepflux, PO4flux, PO4deepflux, DICflux, DICdeepflux, 
      FDETflux, FDETdeepflux, SDETflux, SDETdeepflux, FePdeepflux, CaPdeepflux, 
      OrgCflux, OrgNflux, OrgPflux}, the sediment-water and burial fluxes, in nmol/cm2/d.
   \item \code{DINDIPflux, DINDIPmean, DINDIPdeep}, the dissolved nitrogen to phosphorus ratio of flux, sediment concentrations and deep (burial) concentration.
   \item \code{TotMin, TotOxic, TotDenit, TotAnoxic}, total mineralisation, total oxic mineralisation, denitrification and anoxic mineralisation, in nmol/cm2/d.
   \item \code{PartOxic, PartDenit, PartAnoxic}, the fraction of mineralisation due to oxic, denitrification and anoxic mineralisation.
   \item \code{TotNitri, TotODUoxid, TotFePprod, TotCaPprod, TotFePdesorp, TotCaPdiss, TotNprod, TotPprod, TotNH3ads}, integrated rates, nmol/cm2/d.
   \item \code{PartPremoved, PartNremoved},  the total P and N removed, relative to its production.
   \item \code{TOC}, the Total organic carbon concentration profile, \%.
   \item \code{Cprod,Nprod,Pprod,Oxicmin,Denitrific,anoxicmin,nitri,oduox,odudepo,FePadsorp,CaPprod}, rate profiles, nm/cm3 liquid/d.
   \item \code{FePdesorp,CaPdiss}, rate profiles, nm/cm3 solid/d.
  }
}

\author{
  Karline Soetaert
}


\details{
  To solve the model, a steady-state solver from package rootSolve
  (here we used \code{\link[rootSolve]{steady.1D}}) is
  used. 

  For the CNPDIA model, the parameters and their meaning are the following (with default values):
  \itemize{
    \item Cflux       ,  5.00e+02 ,  nmolC/cm2/d    ,                total organic C deposition
    \item pFast       ,  9.00e-01 ,            -    ,                  part FDET in carbon flux
    \item FePflux     ,  0.00e+00 ,  nmolP/cm2/d    ,                    deposition rate of FeP
    \item CaPflux     ,  0.00e+00 ,  nmolP/cm2/d    ,                    deposition rate of CaP
    \item rFast       ,  6.85e-02 ,           /d    ,                           decay rate FDET
    \item rSlow       ,  1.37e-04  ,          /d    ,                           decay rate SDET
    \item NCrFdet     ,  1.51e-01 ,    molN/molC    ,                             NC ratio FDET
    \item NCrSdet     ,  1.51e-01 ,    molN/molC    ,                             NC ratio SDET
    \item PCrFdet     ,  9.43e-03 ,    molP/molC    ,                             PC ratio FDET
    \item PCrSdet     ,  9.43e-03 ,    molP/molC    ,                             PC ratio SDET
    \item BCupLiq     ,  2.00e+00 ,            -  , upper boundary liq. 1:flux, 2:conc, 3:0-grad
    \item BCdownLiq   ,  3.00e+00 ,            -  , lower boundary liq. 1:flux, 2:conc, 3:0-grad
    \item O2bw        ,  3.00e+02 ,      mmol/m3    , upper boundary O2  -if BC=1: flux, 2:conc
    \item NO3bw       ,  1.00e+01 ,      mmol/m3    , upper boundary NO3 -if BC=1: flux, 2:conc
    \item NO2bw       ,  0.00e+00 ,      mmol/m3    , upper boundary NO2 -if BC=1: flux, 2:conc
    \item NH3bw       ,  1.00e+00 ,      mmol/m3    , upper boundary NH3 -if BC=1: flux, 2:conc
    \item ODUbw       ,  0.00e+00 ,      mmol/m3    , upper boundary ODU -if BC=1: flux, 2:conc
    \item PO4bw       ,  5.00e-01 ,      mmol/m3    , upper boundary PO4 -if BC=1: flux, 2:conc
    \item DICbw       ,  2.20e+03 ,      mmol/m3    , upper boundary DIC -if BC=1: flux, 2:conc
    \item O2dw        ,        NA ,      mmol/m3    , lower boundary O2  -if BC=1: flux, 2:conc
    \item NO3dw       ,        NA ,      mmol/m3    , lower boundary NO3 -if BC=1: flux, 2:conc
    \item NO2dw       ,        NA ,      mmol/m3    , lower boundary NO2 -if BC=1: flux, 2:conc
    \item NH3dw       ,        NA ,      mmol/m3    , lower boundary NH3 -if BC=1: flux, 2:conc
    \item ODUdw       ,        NA ,      mmol/m3    , lower boundary ODU -if BC=1: flux, 2:conc
    \item PO4dw       ,        NA ,      mmol/m3    , lower boundary PO4 -if BC=1: flux, 2:conc
    \item DICdw       ,        NA ,      mmol/m3    , lower boundary DIC -if BC=1: flux, 2:conc
    \item w           ,  2.74e-07 ,         cm/d    ,                            advection rate
    \item biot        ,  2.74e-03 ,        cm2/d    ,                  bioturbation coefficient
    \item biotdepth   ,  5.00e+00 ,           cm    ,                      depth of mixed layer
    \item biotatt     ,  1.00e+00 ,           cm    ,         attenuation coeff below biotdepth
    \item irr         ,  0.00e+00 ,           /d    ,                       bio-irrigation rate
    \item irrdepth    ,  5.00e+00 ,           cm    ,                  depth of irrigated layer
    \item irratt      ,  1.00e+00 ,           cm    ,          attenuation coeff below irrdepth
    \item gasflux     ,  0.00e+00 ,         cm/d    ,             piston velocity for dry flats
    \item NH3Ads      ,  1.30e+00 ,            -    ,                 Adsorption coeff ammonium
    \item rnitri1     ,  2.00e+01 ,           /d    ,      Max nitrification rate step1 (NH3ox)
    \item rnitri2     ,  2.00e+01 ,           /d    ,      Max nitrification rate step2 (NO2ox)
    \item ksO2nitri   ,  1.00e+00 ,    mmolO2/m3    ,              half-sat O2 in nitrification
    \item ranammox    ,  1.00e-01 ,           /d    ,                              Anammox rate
    \item ksNO2anammox,  1.00e-01 ,     mmolN/m3    ,                   half-sat NO2 in anammox
    \item rODUox      ,  2.00e+01 ,           /d    ,       Max rate ODU oxidation in one layer
    \item rSurfODUox  ,  0.00e+00 ,           /d    ,         Max rate ODU oxidation with BW O2
    \item ODUoxdepth  ,  5.00e+00 ,           cm    ,        Max depth ODU oxidation with BW O2
    \item ODUoxatt    ,  1.00e+00 ,           cm    ,           depth attenuation ODU oxidation
    \item ksO2oduox   ,  1.00e+00 ,    mmolO2/m3    ,           half-sat O2 in oxidation of ODU
    \item ksO2oxic    ,  3.00e+00 ,    mmolO2/m3    ,        half-sat O2 in oxic mineralisation
    \item ksNO3denit  ,  3.00e+01 ,   mmolNO3/m3    ,           half-sat NO3 in denitrification
    \item kinO2denit  ,  1.00e+00 ,    mmolO2/m3    ,         half-sat O2 inhib denitrification
    \item kinNO3anox  ,  1.00e+00 ,   mmolNO3/m3    ,            half-sat NO3 inhib anoxic degr
    \item kinO2anox   ,  1.00e+00 ,    mmolO2/m3    ,              half-sat O2 inhib anoxic min
    \item pdepo       ,        NA ,            -    ,  part ODU prod lost (NA:estimated from w)
    \item rdepo       ,  0.00e+00 ,           /d    ,                          ODU removal rate
    \item temperature ,  1.00e+01 ,          dgC    ,                               temperature
    \item salinity    ,  3.50e+01 ,          psu    ,                                  salinity
    \item TOC0        ,  5.00e-01 ,            %    ,                    refractory Carbon conc
    \item rFePdesorp  ,  1.00e-02 ,           /d    ,                       rate FeP desorption
    \item rFePadsorp  ,  3.00e-01 ,           /d    ,                       rate FeP adsorption
    \item rCaPprod    ,  0.00e+00 ,           /d    ,                       rate CaP production
    \item rCaPdiss    ,  0.00e+00 ,           /d    ,                      rate CaP dissolution
    \item CPrCaP      ,  2.87e-01 ,    molC/molP    ,                           C:Pratio in CaP
    \item rPads       ,  0.00e+00 ,           /d    ,                       adsorption rate PO4
    \item rPdes       ,  0.00e+00 ,           /d    ,             desorption rate of adsorbed P
    \item maxPads     ,  1.00e+03 ,mmolP/m3solid    ,              Max adsorbed P concentration
    \item por0        ,  9.00e-01 ,            -    ,                          surface porosity
    \item pordeep     ,  5.00e-01 ,            -    ,                             deep porosity
    \item porcoeff    ,  3.00e-01 ,           cm    ,                porosity decay coefficient
    \item formationtype, 1.00e+00 ,            - , formationfactor, 1=sand,2=fine sand,3=general
    \item dilution    ,  0.00e+00 ,           /d    ,       relaxation towards background conc 
    \item Hwater      ,  1.00e+01 ,           cm    ,                 height of water over core
    \item Cfall       ,  1.00e+02 ,         cm/d    ,      fall speed of organic C (FDET, SDET)
    \item FePfall     ,  1.00e+02 ,         cm/d    ,                         fall speed of FeP
    \item CaPfall     ,  1.00e+02 ,         cm/d    ,                         fall speed of CaP
    \item MPBprod     ,  0.00e+00 ,    mmol/m3/d    ,               maximal MPB production rate
    \item kdSed       ,  2.00e+01 ,          /cm    ,              light extinction coefficient
    \item kNO3upt     ,  3.00e+00 ,     mmolN/m3    ,               NO3 limitation constant MPB
    \item kNH3upt     ,  3.00e+00 ,     mmolN/m3    ,               NH3 limitation constant MPB
    \item kPO4upt     ,  1.00e-01 ,     mmolP/m3    ,                 P limitation constant MPB
    \item kDICupt     ,  1.00e+00 ,     mmolC/m3    ,                 C limitation constant MPB
  }
  
and for \code{MPBDIA}, in addition:
  \itemize{
    \item MPBflux     ,  0.00e+00 ,  nmolC/cm2/day  ,                    deposition of MPB from the water
    \item light       ,  1.00e+02 ,    uEinst/m2/s  ,                          light intensity at surface
    \item kdChl       ,  2.00e-04 , /(mgChl/m3)/cm  ,                 Chl-specific extinction coefficient
    \item QmaxCHLN    ,  3.80e+00 ,    mgChl:mmolN  ,                          maximum Chl:N ratio of MPB
    \item QmaxNC      ,  2.00e-01 ,    mmolC:mmolN  ,                            maximum N:C ratio of MPB
    \item QminNC      ,  5.00e-02 ,    mmolC:mmolN  ,                            minimum N:C ratio of MPB
    \item sigmaPSII   ,  3.31e+00 ,  m2/umolQuanta  ,   absorption cross-section of Photosystem II of MPB
    \item kdH         ,  4.50e-08 ,              -  ,          dimensionless damage rate constant of PSII
    \item tau         ,  6.48e-08 ,              d ,turnover time of the electron transport chain of PSII
    \item kr          ,  2.25e+01 ,             /d  ,                          Rate of PSII damage repair
    \item phiM        ,  5.20e-05 ,mmolC/umolQuant  ,             Maximum quantum yield of photosynthesis
    \item a           ,  4.20e-02 ,       m2/mgChl  ,Spectrally-integrated Chla-specific absorption coeff
    \item maxUpNO3    ,  2.00e-01 ,  mmolN/mmolC/d  ,                   maximal uptake rate of NO3 by MPB
    \item maxUpNH3    ,  2.00e-01 ,  mmolN/mmolC/d  ,                   maximal uptake rate of NH3 by MPB
    \item respMPB     ,  1.00e-01 ,             /d  ,                       basal respiration rate of MPB
    \item rG          ,  2.50e-01 ,    mmolC/mmolC  ,                                      Cost of growth
    \item rVNO3       ,  3.87e-01 ,    mmolC/mmolN  ,                              Cost of nitrate uptake
    \item rVN         ,  1.98e-01 ,    mmolC/mmolN  ,                         Cost of ammonium/don uptake
    \item rRNO3       ,  1.98e+00 ,    mmolC/mmolN  ,                           Cost of nitrate reduction
    \item kEps        ,  2.00e-01 ,              -  , fraction of photosynthesis exudated as Carbohydrate
    \item maxGrazing  ,  1.00e-01 ,             /d  ,            mortality rate of MPB induced by grazing
    \item deepMort    ,  1.00e-02 ,             /d  ,                mortality rate of MPB in anoxic zone
    \item rEPS        ,  2.60e-01 ,             /d  ,                        rate of EPS remineralization
    \item MPBfall     ,  1.00e-02 ,           cm/d  ,                       sinking rate of suspended MPB
    \item kdWater     ,  6.00e-05 ,            /cm  ,                  extinction coeff of light in water
}
}


\examples{

#===========================================
# Show parameter values
#===========================================

  MPBDIAparms()
  par(mar = c(3,3,3,3))

#===========================================
# Different carbon fluxes
#===========================================

  out  <- MPBDIAsolve()
  out2 <- MPBDIAsolve(parms = list(Cflux = 200*1e5/12/365))
  out3 <- MPBDIAsolve(parms = list(Cflux = 2*1e5/12/365))
  plot(out, out2, out3, xyswap = TRUE, grid = CNPDIAdepth(out), 
    ylim = c(20,0), mfrow = c(3, 4))

  CNPDIAbudgetO2(out)  
  
#===========================================  
# Different light regimes
#===========================================
 
# start with simple diagenesis 
  out2  <- CNPDIAsolve()

# add rough initial conditions for MPB dynamics  
  yini  <- c(out2$y, rep(1, 500))

#   
  out2a <- MPBDIAsolve(parms = list(light = 0), yini = yini)
  out2b <- MPBDIAsolve(parms = list(light = 10), yini = yini)
  out2c <- MPBDIAsolve(parms = list(light = 100), yini = yini)
  plot(out2b, out2c, which = c("O2", "CHL"), 
     ylim = list(c(1,0), c(10,0)))

  plot(out2b, out2c, which = c("NH3","ODU"), ylim = c(10,0))

}

\references{
  Soetaert K, PMJ Herman and JJ Middelburg, 1996a.
  A model of early diagenetic processes from the shelf to abyssal depths.
  Geochimica Cosmochimica Acta, 60(6):1019-1040.

  Soetaert K, PMJ Herman and JJ Middelburg, 1996b.
  Dynamic response of deep-sea sediments to seasonal variation: a model.
  Limnol. Oceanogr. 41(8): 1651-1668.
}

\keyword{ utilities }
